+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow to provide the description of the service in the Microsoft Excel add-in.
 - `skip_update_for` option is now allowing dot notation.
 - Allow to provide a custom response converter when using as a python module.
+- Allow to set NTLM authentication with current credentials in the Microsoft Excel add-in.
 
 ### Fixed
 - Allow flexible dependencies versions (only requires major component to match).